refactor(day21): remove unnecessary uint casts in tests
Remove explicit uint() casts from test expected values. Go's type
inference automatically converts untyped integer constants to the
correct type based on the solver's return type.

+++ day21_test.go
@@ -3,11 +3,11 @@ package adventofcode2018
 import "testing"
 
 func TestDay21Part1(t *testing.T) {
-	testWithParserBytes(t, 21, file, true, NewDay21, Day21, uint(7216956))
+	testWithParserBytes(t, 21, file, true, NewDay21, Day21, 7216956)
 }
 
 func TestDay21Part2(t *testing.T) {
-	testWithParserBytes(t, 21, file, false, NewDay21, Day21, uint(14596916))
+	testWithParserBytes(t, 21, file, false, NewDay21, Day21, 14596916)
 }
 
 func BenchmarkDay21Part1(b *testing.B) {
